web-app/sai-katuri/requirements.txt
phidata
openai
python-dotenv
packaging
uvicorn
groq
gradio
